refactor: Move MyWay components
